+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Version 0.31
 Currently in development.
 
 - Added federated login
+- Added OpenID Connect (OIDC) login
 - Added setting for disabling object assignments for individual locations
 - Refactored related objects tree to improve performance
 - Allow notes on arrays in schema